Refactor: skipif for Windows 2/n (#6268)
* win

* isort

* flake8
