+++ .github/workflows/events-recurrent.yml
@@ -1,4 +1,4 @@
-name: Recurent events
+name: Recurrent events
 
 # https://jasonet.co/posts/scheduled-actions/
 # https://github.community/t/distinct-job-for-each-schedule/17811/2
