+++ pytorch_lightning/profiler/__init__.py
@@ -194,12 +194,7 @@ To visualize the profiled operation, you can either:
 
 """
 
-from pytorch_lightning.profiler.profilers import (
-    AdvancedProfiler,
-    BaseProfiler,
-    PassThroughProfiler,
-    SimpleProfiler,
-)
+from pytorch_lightning.profiler.profilers import AdvancedProfiler, BaseProfiler, PassThroughProfiler, SimpleProfiler
 from pytorch_lightning.profiler.pytorch import PyTorchProfiler
 
 __all__ = [
