+++ tests/accelerators/test_ddp.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-import platform
 from unittest.mock import patch
 
 import pytest
@@ -21,7 +20,7 @@ import torch
 from pytorch_lightning import Trainer
 from tests.accelerators import ddp_model, DDPLauncher
 from tests.helpers.boring_model import BoringModel
-from tests.helpers.skipif import RunIf
+from tests.helpers.runif import RunIf
 from tests.utilities.distributed import call_training_script
 
 CLI_ARGS = '--max_epochs 1 --gpus 2 --accelerator ddp'
@@ -83,7 +82,7 @@ def test_cli_to_pass(tmpdir, args=None):
     return '1'
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Distributed training is not supported on Windows")
+@RunIf(skip_windows=True)
 @pytest.mark.skipif(torch.cuda.is_available(), reason="test doesn't requires GPU machine")
 def test_torch_distributed_backend_env_variables(tmpdir):
     """
