+++ tests/accelerators/test_dp.py
@@ -21,8 +21,8 @@ from pytorch_lightning.callbacks import EarlyStopping
 from pytorch_lightning.core import memory
 from tests.helpers import BoringModel
 from tests.helpers.datamodules import ClassifDataModule
+from tests.helpers.runif import RunIf
 from tests.helpers.simple_models import ClassificationModel
-from tests.helpers.skipif import RunIf
 
 PRETEND_N_OF_GPUS = 16
