+++ tests/callbacks/test_gpu_stats_monitor.py
@@ -24,7 +24,7 @@ from pytorch_lightning.loggers.csv_logs import ExperimentWriter
 from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.helpers import BoringModel
-from tests.helpers.skipif import RunIf
+from tests.helpers.runif import RunIf
 
 
 @RunIf(min_gpus=1)
