+++ tests/callbacks/test_quantization.py
@@ -21,8 +21,8 @@ from pytorch_lightning.callbacks import QuantizationAwareTraining
 from pytorch_lightning.metrics.functional.mean_relative_error import mean_relative_error
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.helpers.datamodules import RegressDataModule
+from tests.helpers.runif import RunIf
 from tests.helpers.simple_models import RegressionModel
-from tests.helpers.skipif import RunIf
 
 
 @pytest.mark.parametrize("observe", ['average', pytest.param('histogram', marks=RunIf(min_torch="1.5"))])
