+++ tests/core/test_metric_result_integration.py
@@ -11,9 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import sys
 
-import pytest
 import torch
 import torch.distributed as dist
 import torch.multiprocessing as mp
@@ -21,6 +19,7 @@ import torch.multiprocessing as mp
 import tests.helpers.utils as tutils
 from pytorch_lightning.core.step_result import Result
 from pytorch_lightning.metrics import Metric
+from tests.helpers.runif import RunIf
 
 
 class DummyMetric(Metric):
@@ -90,7 +89,7 @@ def _ddp_test_fn(rank, worldsize):
             assert epoch_expected[k] == epoch_log[k]
 
 
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(skip_windows=True)
 def test_result_reduce_ddp():
     """Make sure result logging works with DDP"""
     tutils.reset_seed()
