+++ tests/core/test_results.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import random
-import sys
 from pathlib import Path
 
 import pytest
@@ -26,7 +25,7 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.core.step_result import Result
 from pytorch_lightning.trainer.states import TrainerState
 from tests.helpers import BoringDataModule, BoringModel
-from tests.helpers.skipif import RunIf
+from tests.helpers.runif import RunIf
 
 
 def _setup_ddp(rank, worldsize):
@@ -49,7 +48,7 @@ def _ddp_test_fn(rank, worldsize, result_cls: Result):
 
 
 @pytest.mark.parametrize("result_cls", [Result])
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(skip_windows=True)
 def test_result_reduce_ddp(result_cls):
     """Make sure result logging works with DDP"""
     tutils.reset_seed()
