+++ tests/metrics/test_ddp.py
@@ -1,9 +1,8 @@
-import sys
-
 import pytest
 import torch
 
 from pytorch_lightning.metrics import Metric
+from tests.helpers.runif import RunIf
 from tests.metrics.test_metric import Dummy
 from tests.metrics.utils import setup_ddp
 
@@ -40,7 +39,7 @@ def _test_ddp_sum_cat(rank, worldsize):
     assert dummy.bar == worldsize
 
 
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(skip_windows=True)
 @pytest.mark.parametrize("process", [_test_ddp_cat, _test_ddp_sum, _test_ddp_sum_cat])
 def test_ddp(process):
     torch.multiprocessing.spawn(process, args=(2, ), nprocs=2)
@@ -66,7 +65,7 @@ def _test_non_contiguous_tensors(rank, worldsize):
     metric.update(torch.randn(10, 5)[:, 0])
 
 
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(skip_windows=True)
 def test_non_contiguous_tensors():
     """ Test that gather_all operation works for non contiguous tensors """
     torch.multiprocessing.spawn(_test_non_contiguous_tensors, args=(2, ), nprocs=2)
