+++ tests/trainer/logging_/test_train_loop_logging_1_0.py
@@ -31,7 +31,7 @@ from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
 from pytorch_lightning.core.lightning import LightningModule
 from tests.helpers.boring_model import BoringModel, RandomDictDataset, RandomDictStringDataset
 from tests.helpers.deterministic_model import DeterministicModel
-from tests.helpers.skipif import RunIf
+from tests.helpers.runif import RunIf
 
 
 @mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
