+++ tests/trainer/properties/test_get_model.py
@@ -11,14 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import sys
-
-import pytest
 
 from pytorch_lightning import Trainer
 from tests.accelerators import DDPLauncher
 from tests.helpers.boring_model import BoringModel
-from tests.helpers.skipif import RunIf
+from tests.helpers.runif import RunIf
 
 
 class TrainerGetModel(BoringModel):
@@ -47,7 +44,7 @@ def test_get_model(tmpdir):
     trainer.fit(model)
 
 
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(skip_windows=True)
 def test_get_model_ddp_cpu(tmpdir):
     """
     Tests that `trainer.lightning_module` extracts the model correctly when using ddp on cpu
@@ -86,8 +83,7 @@ def test_get_model_gpu(tmpdir):
     trainer.fit(model)
 
 
-@RunIf(min_gpus=1)
-@pytest.mark.skipif(sys.platform == "win32", reason="DDP not available on windows")
+@RunIf(min_gpus=1, skip_windows=True)
 @DDPLauncher.run("--accelerator [accelerator]", max_epochs=["1"], accelerator=["ddp", "ddp_spawn"])
 def test_get_model_ddp_gpu(tmpdir, args=None):
     """
